code speedup
setting this to float32 should speed up the np.cumsum step
inside the models, which is a big bottleneck

+++ automated_forecasting/phenology/apply_phenology_models.py
@@ -66,7 +66,7 @@ def run():
         for climate in current_climate_forecasts:
             doy_series =  pd.TimedeltaIndex(climate.time.values - doy_0, freq='D').days.values
             
-            species_ensemble.append(model.predict(predictors={'temperature': climate.tmean.values,
+            species_ensemble.append(model.predict(predictors={'temperature': climate.tmean.values.astype(np.float32),
                                                               'doy_series' : doy_series}))
         
         species_ensemble = np.array(species_ensemble).astype(np.float)
